# openapi/api-v1.yaml
{
  "openapi": "3.1.0",
  "info": {
    "title": "RPP AI Mobile API",
    "version": "1.0.0-target",
    "summary": "Kontrak target API v1 untuk Laravel dan Android",
    "description": "Source of truth implementasi mendatang. Dokumentasi ini tidak menyatakan endpoint telah tersedia pada runtime saat ini."
  },
  "servers": [
    {
      "url": "{baseUrl}/api/v1",
      "variables": {
        "baseUrl": {
          "default": "http://localhost:8000"
        }
      }
    }
  ],
  "security": [
    {
      "sanctum": []
    }
  ],
...
  "paths": {
...
    "/settings": {
      "get": {
        "tags": [
          "Settings"
        ],
        "operationId": "getSettings",
        "summary": "Pengaturan sekolah",
        "responses": {
          "200": {
            "description": "Berhasil",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SuccessEnvelope"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/Settings"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/E401"
          },
          "403": {
            "$ref": "#/components/responses/E403"
          },
          "404": {
            "$ref": "#/components/responses/E404"
          },
          "422": {
            "$ref": "#/components/responses/E422"
          },
          "429": {
            "$ref": "#/components/responses/E429"
          },
          "500": {
            "$ref": "#/components/responses/E500"
          }
        },
        "security": [
          {
            "sanctum": []
          }
        ]
      },
      "patch": {
        "tags": [
          "Settings"
        ],
        "operationId": "updateSettings",
        "summary": "Perbarui pengaturan sekolah (capability settings.manage)",
        "responses": {
          "200": {
            "description": "Berhasil",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SuccessEnvelope"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/Settings"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/E401"
          },
          "403": {
            "$ref": "#/components/responses/E403"
          },
          "404": {
            "$ref": "#/components/responses/E404"
          },
          "422": {
            "$ref": "#/components/responses/E422"
          },
          "429": {
            "$ref": "#/components/responses/E429"
          },
          "500": {
            "$ref": "#/components/responses/E500"
          }
        },
        "security": [
          {
            "sanctum": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/SettingsInput"
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "sanctum": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "Sanctum"
      }
    },
...
    "responses": {
      "E401": {
        "description": "Tidak terautentikasi",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "E403": {
        "description": "Tidak diizinkan",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "E404": {
        "description": "Tidak ditemukan",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
...
      "E422": {
        "description": "Validasi gagal",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "E429": {
        "description": "Terlalu banyak permintaan",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "E500": {
        "description": "Kesalahan server",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
...
    },
    "schemas": {
      "SuccessEnvelope": {
        "type": "object",
        "required": [
          "data"
        ],
        "properties": {
          "data": {},
          "meta": {
            "$ref": "#/components/schemas/PaginationMeta"
          },
          "links": {
            "$ref": "#/components/schemas/PaginationLinks"
          }
        }
      },
      "Error": {
        "type": "object",
        "required": [
          "message",
          "code"
        ],
        "properties": {
          "message": {
            "type": "string"
          },
          "code": {
            "type": "string"
          },
          "errors": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        }
      },
      "PaginationMeta": {
        "type": "object",
        "properties": {
          "current_page": {
            "type": "integer"
          },
          "last_page": {
            "type": "integer"
          },
          "per_page": {
            "type": "integer"
          },
          "total": {
            "type": "integer"
          },
          "from": {
            "type": "integer"
          },
          "to": {
            "type": "integer"
          }
        }
      },
      "PaginationLinks": {
        "type": "object",
        "properties": {
          "first": {
            "type": [
              "string",
              "null"
            ],
            "format": "uri"
          },
          "last": {
            "type": [
              "string",
              "null"
            ],
            "format": "uri"
          },
          "prev": {
            "type": [
              "string",
              "null"
            ],
            "format": "uri"
          },
          "next": {
            "type": [
              "string",
              "null"
            ],
            "format": "uri"
          }
        }
      },
...
      "Settings": {
        "type": "object",
        "properties": {
          "nama_sekolah": {
            "type": [
              "string",
              "null"
            ]
          },
          "nsm": {
            "type": [
              "string",
              "null"
            ]
          },
          "npsn": {
            "type": [
              "string",
              "null"
            ]
          },
          "alamat": {
            "type": [
              "string",
              "null"
            ]
          },
          "logo_url": {
            "type": [
              "string",
              "null"
            ],
            "format": "uri"
          },
          "logo_kanan_url": {
            "type": [
              "string",
              "null"
            ],
            "format": "uri"
          },
          "kop_surat_url": {
            "type": [
              "string",
              "null"
            ],
            "format": "uri"
          }
        }
      },
      "SettingsInput": {
        "type": "object",
        "properties": {
          "nama_sekolah": {
            "type": [
              "string",
              "null"
            ],
            "maxLength": 255
          },
          "nsm": {
            "type": [
              "string",
              "null"
            ],
            "maxLength": 20
          },
          "npsn": {
            "type": [
              "string",
              "null"
            ],
            "maxLength": 20
          },
          "alamat": {
            "type": [
              "string",
              "null"
            ],
            "maxLength": 500
          },
          "logo": {
            "type": [
              "string",
              "null"
            ],
            "format": "binary"
          },
          "logo_kanan": {
            "type": [
              "string",
              "null"
            ],
            "format": "binary"
          },
          "kop_surat": {
            "type": [
              "string",
              "null"
            ],
            "format": "binary"
          }
        }
      },
...
    }
  }
}